+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - Updated minimum required version to v4.0.2
+- Fixed `run_in_transaction!` to rollback when interaction finished with errors
 
 ## [1.0.1] - 2021-05-13